requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
numpy==1.26.1
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.40
psycopg2-binary==2.9.9
PyPDF2==3.0.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
reportlab==4.0.7
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
